Revert "ARM: update ctf to 380.68, enable pptp and l2tp ctf"
This reverts commit 1c0b84d1f77a8a0e5fd3e6f905eec8148153c1c8.

+++ release/src-rt/platform.mak
@@ -81,10 +81,18 @@ define platformKernelConfig
 	if [ "$(ARM)" = "y" ]; then \
 		mkdir -p $(SRCBASE)/router/ctf_arm/linux; \
 		if [ "$(ARMCPUSMP)" = "up" ]; then \
-			cp -f $(SRCBASE)/router/ctf_arm/bcm6_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+			if [ "$(CTF_PPTP_L2TP)" = "y" ]; then \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm6_up_pptp_l2tp/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+			else \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm6_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+			fi; \
 			cp -f $(SRCBASE)/router/ufsd/broadcom_arm_up/ufsd.ko.46_up $(SRCBASE)/router/ufsd/broadcom_arm/ufsd.ko; \
 		else \
-			cp -f $(SRCBASE)/router/ctf_arm/bcm6/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+			if [ "$(CTF_PPTP_L2TP)" = "y" ]; then \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm6_pptp_l2tp/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+			else \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm6/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+			fi; \
 		fi; \
 		if [ "$(DPSTA)" = "y" ]; then \
 			mkdir -p $(SRCBASE)/router/dpsta/linux; \
